Refactor : 회원 정보 입력 로직 개선
- auth/info >> auth/user
- auth/me >> auth/user/get
- input 유효성 검사
- userInfoMutation 에러 처리
